move LightblueException to the top level
