+++ core/src/test/java/com/redhat/lightblue/client/response/TestDefaultLightblueDataResponse.java
@@ -11,6 +11,7 @@ import org.junit.Test;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.redhat.lightblue.client.LightblueException;
 import com.redhat.lightblue.client.util.JSON;
 
 public class TestDefaultLightblueDataResponse {
