+++ lightblue-client-integration-test/src/main/java/com/redhat/lightblue/client/integration/test/LightblueClientTestHarness.java
@@ -9,9 +9,9 @@ import java.util.Set;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.redhat.lightblue.client.LightblueClientConfiguration;
+import com.redhat.lightblue.client.LightblueException;
 import com.redhat.lightblue.client.http.LightblueHttpClient;
 import com.redhat.lightblue.client.request.metadata.MetadataGetEntityVersionsRequest;
-import com.redhat.lightblue.client.response.LightblueException;
 import com.redhat.lightblue.client.response.LightblueResponse;
 import com.redhat.lightblue.client.test.request.DataInsertRequestStub;
 import com.redhat.lightblue.rest.integration.AbstractCRUDControllerWithRest;
